ExceptionHandler

+++ Antomi.Service/Implementations/CategoryService.cs
@@ -1,6 +1,7 @@
 ﻿using Antomi.Core.Entities;
 using Antomi.Core.IRepositories;
 using Antomi.Service.DTOs.CategoryDTOs;
+using Antomi.Service.Exceptions;
 using Antomi.Service.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -20,6 +21,9 @@ namespace Antomi.Service.Implementations
         }
         public async Task<CategoryGetDto> CreateAsync(CategoryPostDto postDto)
         {
+            if (await categoryRepository.IsExistAsync(x => x.Name.ToUpper() == postDto.Name.ToUpper() && x.isDeleted == false))
+                throw new RecordDublicateException("This Category Already Exist(" + postDto.Name + ")");
+
             Category category = new Category()
             {
                 Name = postDto.Name
@@ -39,7 +43,10 @@ namespace Antomi.Service.Implementations
 
         public async Task Delete(int id)
         {
+            if (!await categoryRepository.IsExistAsync(x => x.Id == id && x.isDeleted == false))
+                throw new ItemNotFoundException("Item Not Found by Id(" +id+ ")");
             Category category = await categoryRepository.GetAsync(x => x.Id == id && x.isDeleted == false);
+
             categoryRepository.Remove(category);
             await categoryRepository.SaveDbAsync();
         }
@@ -47,7 +54,10 @@ namespace Antomi.Service.Implementations
         public async Task<CategoryGetDto> GetByIdAsync(int id)
         {
             Category category = await categoryRepository.GetAsync(x => x.Id == id && x.isDeleted == false);
-
+            if(category==null)
+                throw new ItemNotFoundException("Item Not Found by Id(" + id.ToString() + ")");
+            if (!await categoryRepository.IsExistAsync(x => x.Id == id && x.isDeleted == false))
+                throw new ItemNotFoundException("Item Not Found by Id(" + id.ToString() + ")");
             CategoryGetDto categoryGetDto = new CategoryGetDto()
             {
                 Id = category.Id,
@@ -59,7 +69,10 @@ namespace Antomi.Service.Implementations
 
         public async Task<CategoryGetDto> UpdateAsync(int id, CategoryPostDto postDto)
         {
-
+            if (!await categoryRepository.IsExistAsync(x => x.Id == id && x.isDeleted == false))
+                throw new ItemNotFoundException("Item Not Found by Id(" + id + ")");
+            if (await categoryRepository.IsExistAsync(x => x.Name.ToUpper() == postDto.Name.ToUpper() && x.isDeleted == false))
+                throw new RecordDublicateException("This Category Already Exist(" + postDto.Name + ")");
             Category category = await categoryRepository.GetAsync(x => x.Id == id && x.isDeleted == false);
             category.Name = postDto.Name;
             await categoryRepository.SaveDbAsync();
